+++ src/main/java/com/fashion_shop/controller/ProductController.java
@@ -1,6 +1,7 @@
 package com.fashion_shop.controller;
 
 import com.fashion_shop.model.Product;
+import com.fashion_shop.model.User;
 import com.fashion_shop.model.dto.ResponseDto;
 import com.fashion_shop.service.ImageService;
 import com.fashion_shop.service.ProductService;
@@ -33,8 +34,9 @@ public class ProductController {
     }
 
     @PostMapping
-    ResponseEntity<ResponseDto> create(@RequestBody Product product){
-        if(!ProductValidator.validateCreateProduct(product)){
+    ResponseEntity<ResponseDto> create(@RequestBody Product product,
+                                       @RequestHeader String userId){
+        if(!ProductValidator.validateCreateProduct(product, userId)){
             throw  new ResponseStatusException(HttpStatus.BAD_REQUEST,
                     "user data is invalid to create product");
         }
@@ -45,8 +47,10 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<ResponseDto> update(@PathVariable Long id, @RequestBody Product product){
-        if (!ProductValidator.validateUpdateProduct(product)) {
+    ResponseEntity<ResponseDto> update(@PathVariable Long id,
+                                       @RequestBody Product product,
+                                       @RequestHeader String userId){
+        if (!ProductValidator.validateUpdateProduct(product, userId)) {
             throw new ResponseStatusException(
                     HttpStatus.BAD_REQUEST,
                     "user data is invalid to update product with id:" + id
@@ -59,7 +63,14 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<ResponseDto> delete(@PathVariable long id){
+    ResponseEntity<ResponseDto> delete(@PathVariable long id, @RequestHeader String userId){
+        if (!ProductValidator.validateDeleteProduct(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "user is unauthorized, please sign in first:"
+            );
+        }
+        imageService.delete(id);
         productService.delete(id);
         ResponseDto responseDto = new ResponseDto("Product deleted.");
         responseDto.addInfo("productId", String.valueOf(id));
